introduce Reactor::ResourceActionable
as a way to DRY up controller action_event declarations
while also enforcing a nice naming convention for your activity model.

Consider the benefits of a event data model that is a superset of the database model.
And in an even more cooler world, the data model will match the business logic.
And thus we have potential for parity between our business, database, and event DSL