Fix bug where delay_for breaks Sidekiq workers
In version 3.5.2 of Sidekiq, calling delay_for on a Sidekiq worker class began raising an exception.
A specific usage of a Subscriber (on_event :my_event, :my_method) which was also a Worker would therefore break.
This hacks it to call perform_in instead of delay_for in this situation.
It also defines perform, which calls the passed class method and creates the appropriate event from the passed data.

Updated the specs a bit for clarity.

Removed suppression of an unrelated and obsolete deprecation warning.