Remove .delay from generic event calls
This required creating a separate worker for database subscribers. Looks like we were always firing database subscribers, even in test mode. The worker no longer does this, so had to add enabling that in a few places in the specs. We had one expectation on the `DelayedExtensions` worker, which has been changed to an expect for `perform_at`